CRITICAL BEATMAP FIX
fixed the awful delay on beatmaps stored in files, since the old method used to delay the the notes by a frame, and with a lot of notes it makes the game go horribly off-sync. plus the delay wasant even consistant so maps wouldnt be able to circumvent this. well its fixed now

other things added
+ Accent colored score
+ X to view PP mid-play
+ Start and Setting Button Recolor
- Map 2 Removed
- Map 4 Removed

PS to mappers: Currently you cannot exceed 500 objects in your level. i might make it 1000 if its necessary